+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Fix CI (PR #418).
 - Add continuous benchmarking using bencher (PR #423, PR #425).
+- Handle unnamed `SingleCellExperiment` assays in `as_AnnData()` by automatically assigning names with a warning (PR #420).
 - Allow manually setting chunk size for HDF5 writes (PR #424).
 - Add auto-chunking for HDF5 writes to improve performance (PR #424).